+++ pkg/kubewarden/tsconfig.json
@@ -20,8 +20,7 @@
       "./types"
     ],
     "types": [
-      "node",
-      "webpack-env",
+      "@types/node",
       "@types/lodash",
       "rancher",
       "shell",
