+++ pkg/kubewarden/types.ts
@@ -1,44 +1,3 @@
-export const KUBEWARDEN_PRODUCT_NAME = 'kubewarden';
-export const KUBEWARDEN_PRODUCT_GROUP = 'policies.kubewarden.io';
-
-export const CHART_NAME = 'rancher-kubewarden';
-
-export const KUBEWARDEN_DASHBOARD = 'dashboard';
-export const KUBEWARDEN_REPO = 'https://charts.kubewarden.io';
-
-export const KUBEWARDEN_CHARTS = {
-  CONTROLLER:       'kubewarden-controller',
-  DEFAULTS:         'kubewarden-defaults',
-};
-
-export const KUBEWARDEN_APPS = {
-  RANCHER_CONTROLLER: 'rancher-kubewarden-controller',
-  RANCHER_DEFAULTS:   'rancher-kubewarden-defaults'
-};
-
-export const KUBEWARDEN = {
-  ADMISSION_POLICY:         'policies.kubewarden.io.admissionpolicy',
-  CLUSTER_ADMISSION_POLICY: 'policies.kubewarden.io.clusteradmissionpolicy',
-  POLICY_SERVER:            'policies.kubewarden.io.policyserver'
-};
-
-export const METRICS_DASHBOARD = {
-  POLICY_SERVER: 'kubewarden-dashboard-policyserver',
-  POLICY:        'kubewarden-dashboard-policy'
-};
-
-export const VALUES_STATE = {
-  FORM: 'FORM',
-  YAML: 'YAML',
-};
-
-export const YAML_OPTIONS = [
-  {
-    labelKey: 'catalog.install.section.chartOptions',
-    value:    VALUES_STATE.FORM,
-  },
-  {
-    labelKey: 'catalog.install.section.valuesYaml',
-    value:    VALUES_STATE.YAML,
-  }
-];
+export * from './types/kubewarden';
+export * from './types/policy';
+export * from './types/grafana';
